+++ app/Http/Controllers/HomeController.php
@@ -10,9 +10,4 @@ class HomeController extends Controller
     {
         return view('home');
     }
-
-    public function about()
-    {
-        return view('about');
-    }
 }