+++ resources/views/about.blade.php
@@ -1,3 +1,3 @@
-<x-layout>
-    <h1>Welkom op de About Me pagina</h1>
-</x-layout>
+<x-main pagename="About Us">
+<h1>Over Ons</h1>
+</x-main>