+++ resources/views/legosets/index.blade.php
@@ -1,8 +1,0 @@
-<div class="container">
-    <h1>Legosets</h1>
-    <ul>
-        @foreach($sets as $set)
-            <li>{{ $set->name }}</li>
-        @endforeach
-    </ul>
-</div>