+++ resources/views/sets/index.blade.php
@@ -1,0 +1,3 @@
+<x-main pagename="LEGOsets">
+<h1>LEGOsets</h1>
+</x-main>