+++ routes/web.php
@@ -1,13 +1,16 @@
 <?php
 
+use App\Http\Controllers\AboutController;
 use App\Http\Controllers\HomeController;
-use App\Http\Controllers\LegosetController;
+use App\Http\Controllers\SetsController;
 use App\Http\Controllers\ProfileController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', [HomeController::class, 'index']);
 
-Route::get('/about', [HomeController::class, 'about']);
+Route::get('/about', [AboutController::class, 'about']);
+
+Route::resource('sets', SetsController::class);
 
 Route::get('/dashboard', function () {
     return view('dashboard');
@@ -20,5 +23,3 @@ Route::middleware('auth')->group(function () {
 });
 
 require __DIR__.'/auth.php';
-
-Route::resource('legosets', LegosetController::class);